bring back b2b and b2c configuration modules
